moved stuff out of ccc.py and into ccc_utils.py
